pyproject.toml
# pyproject.toml - Enhanced Python configuration

[project]
name = "agentflow"
version = "0.1.0"
description = "AgentFlow MVP scaffold"
requires-python = ">=3.10"
dependencies = [
    "alembic==1.16.4",
    "cryptography==45.0.6",
    "email-validator==2.2.0",
    "fastapi==0.115.12",
    "fastapi-guard==0.5.0",
    "httpx==0.28.1",
    "langgraph>=0.1.0",
    "langgraph-checkpoint-postgres>=0.1.0",
    "langgraph-checkpoint-redis>=0.1.0",
    "loguru==0.7.3",
    "mcp[cli]==1.13.0",
    "mem0ai==0.1.116",
    "opentelemetry-api>=1.20",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]==3.2.9",
    "pydantic==2.11.7",
    "pydantic-ai-slim[openai]==0.7.4",
    "pydantic-settings==2.10.1",
    "pyjwt==2.10.1",
    "pyotp==2.9.0",
    "python-dotenv==1.1.1",
    "qdrant-client==1.15.1",
    "r2r==3.6.6",
    "redis==6.4.0",
    "slowapi==0.1.9",
    "SQLAlchemy==2.0.43",
    "tenacity==9.1.2",
    "pybreaker==1.2.0",
    "uvicorn[standard]==0.35.0",
]

[project.optional-dependencies]
dev = [
    "aiosqlite==0.21.0",
    "bandit",
    "black",
    "fakeredis==2.31.0",
    "flake8",
    "hypothesis",
    "mypy>=1.5",
    "pre-commit>=3.0",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-cov",
    "respx==0.22.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["apps", "packages"]

[dependency-groups]
default = []
dev = [
    "aiosqlite==0.21.0",
    "bandit",
    "black",
    "fakeredis==2.31.0",
    "flake8",
    "hypothesis",
    "mypy>=1.5",
    "pre-commit>=3.0",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-cov",
    "respx==0.22.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=apps --cov-report=term-missing"
asyncio_mode = "auto"
